# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API untuk sistem pengelolaan museum",
        "title": "API Pengelolaan Gudang Koleksi Museum",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ghaidafasya24",
            "email": "714220031@std.ulbi.ac.id"
        },
        "version": "1.0"
    },
    "host": "inventorymuseum-de54c3e9b901.herokuapp.com",
    "basePath": "/api",
    "paths": {
...
        "/rak": {
            "get": {
                "description": "Mengambil seluruh data rak dari database MongoDB.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Tempat Penyimpanan"
                ],
                "summary": "Get All Rak",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/model.GetAllRakResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menambahkan data rak baru ke dalam sistem.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Tempat Penyimpanan"
                ],
                "summary": "Insert Rak",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nama Rak",
                        "name": "nama_rak",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Data Rak berhasil ditambahkan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/rak/{id}": {
            "get": {
                "description": "Mengambil data rak berdasarkan ID rak",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Tempat Penyimpanan"
                ],
                "summary": "Get Rak by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Rak",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data rak berhasil ditemukan",
                        "schema": {
                            "$ref": "#/definitions/model.Rak"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memperbarui data rak berdasarkan ID rak",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Tempat Penyimpanan"
                ],
                "summary": "Update Rak",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Rak",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nama Rak",
                        "name": "nama_rak",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data rak berhasil diperbarui",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "Menghapus data rak berdasarkan ID rak",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data Tempat Penyimpanan"
                ],
                "summary": "Delete Rak by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Rak",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data rak berhasil dihapus",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.GetAllRakResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RakResponseItem"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Berhasil mengambil semua data rak"
                },
                "total": {
                    "type": "integer",
                    "example": 12
                }
            }
        },
...
        "model.Rak": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "nama_rak": {
                    "type": "string"
                }
            }
        },
        "model.RakResponseItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "693a3a7a416cd8d592b5058e"
                },
                "nama_rak": {
                    "type": "string",
                    "example": "Rak 2"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer \u003cJWT Token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}